refactor: extract content (folders/links/files) out of courses.py + main.py
Eighth phase — the largest slice of the courses.py/main.py decomposition.
Link and File (previously defined inside app/models/course.py, a hub
file with 6 unrelated concerns) are split out and merged with
ContentFolder (app/models/content_folder.py) into app/content/models.py.

Moved into app/content/routes.py (two blueprints: content_bp for /api
JSON routes, content_files_bp with no prefix for the two file-serving
routes, mirroring the issues_pages/discussion_pages pattern):
- Link CRUD + archive/restore/soft-delete, File CRUD + archive/restore/
  soft-delete (the File lifecycle trio was in main.py's repetitive
  quiz/assignment/file archive-restore-delete block; quiz and assignment
  stay in main.py for a later phase, only file moves here since File is
  content's model)
- Full content-folder tree CRUD, reorder, and move endpoints
- serve_file / serve_question_image (main.py)
- Cross-course material import: api_get_importable_materials +
  api_import_materials (main.py) — moved verbatim rather than refactored
  into per-app "clone" services as originally sketched in the plan, to
  keep this already-large phase behavior-preserving; the service-based
  abstraction is a reasonable follow-up, not required to hit the
  maintainability goal

Verified via grep against app/static/js/ (per the plan's explicit
instruction not to touch these without checking first) which of the
three seemingly-duplicate "move to folder" endpoints are actually used:
- api_move_to_folder (POST .../content/move-to-folder) — used by
  course_detail.js
- api_move_material (POST /api/materials/<type>/<id>/move) — used by
  materials-list-v2.js; its own code comment explains it's
  intentionally type-in-path to avoid ID collisions across Quiz/
  Assignment/File/Link tables
- api_move_material_to_folder (PUT /api/materials/<id>/move) — zero
  references anywhere in the frontend, AND it probes tables in a fixed
  order to guess the material's type, which is exactly the ID-collision
  bug the other endpoint's comment warns about. Deleted as dead code
  rather than "consolidated" — there was nothing live to consolidate it
  with.

Also fixed two gaps from earlier phases while touching these files:
api_create_quiz (course-scoped quiz creation) was still sitting in
courses.py instead of app/quiz/, and api_update_assignment was still in
courses.py instead of app/assignment/ — both per the original
decomposition table. api_update_assignment needed a second blueprint
(assignment_api_bp, url_prefix='/api') since assignment_bp's prefix is
fixed to /assignment.

tests/test_file_upload.py imported `allowed_file` from
app.blueprints.courses directly (bypassing any blueprint abstraction) —
updated to app.content.routes, the only genuine regression caught by
the verification pass (5 failing tests, fixed before commit).

Verified: create_app() now reports 222 routes (223 minus the one dead
duplicate route deleted), SQLAlchemy metadata still has 33 tables, and
pytest is unchanged from baseline (25 pre-existing failures, 99
passing) after fixing the test_file_upload.py import.
