+++ app/models/course.py
@@ -1,6 +1,4 @@
-import datetime
-import enum
-from typing import List, Optional
+from typing import List
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 from app.core.extensions import db
 from app.helpers import get_jakarta_now
@@ -76,61 +74,3 @@ class UserCourseOrder(db.Model):
         return f'<UserCourseOrder User:{self.user_id} Course:{self.course_id} Order:{self.manual_order}>'
 
 
-class Link(db.Model):
-    __tablename__ = 'links'
-
-    id: Mapped[int] = mapped_column(db.Integer, primary_key=True)
-    name: Mapped[str] = mapped_column(db.String(200), nullable=False)
-    url: Mapped[str] = mapped_column(db.String(500), nullable=False)
-    description: Mapped[Optional[str]] = mapped_column(db.Text, nullable=True)
-    course_id: Mapped[int] = mapped_column(db.Integer, db.ForeignKey('courses.id'), nullable=False, index=True)
-
-    # Folder organization
-    folder_id: Mapped[Optional[int]] = mapped_column(db.Integer, db.ForeignKey('content_folders.id'), nullable=True, index=True)
-    order: Mapped[int] = mapped_column(db.Integer, default=0, nullable=False)
-
-    # Archive status
-    is_archived: Mapped[bool] = mapped_column(db.Boolean, default=False, nullable=False, index=True)
-
-    # Trash (Ruang TPS)
-    is_trashed: Mapped[bool] = mapped_column(db.Boolean, default=False, nullable=False, index=True)
-    trashed_at: Mapped[Optional[datetime.datetime]] = mapped_column(db.DateTime, nullable=True)
-
-    created_at: Mapped[datetime.datetime] = mapped_column(db.DateTime, default=get_jakarta_now)
-
-    course: Mapped[Course] = relationship('Course', back_populates='links')
-    folder = relationship('ContentFolder', back_populates='links', foreign_keys=[folder_id])
-
-    def __repr__(self) -> str:
-        return f'<Link {self.name}>'
-
-
-class File(db.Model):
-    __tablename__ = 'files'
-
-    id: Mapped[int] = mapped_column(db.Integer, primary_key=True)
-    name: Mapped[str] = mapped_column(db.String(200), nullable=False)
-    description: Mapped[Optional[str]] = mapped_column(db.Text, nullable=True)
-    filename: Mapped[str] = mapped_column(db.String(200), nullable=False)
-    course_id: Mapped[int] = mapped_column(db.Integer, db.ForeignKey('courses.id'), nullable=False, index=True)
-    start_date: Mapped[Optional[datetime.datetime]] = mapped_column(db.DateTime, nullable=True)
-    end_date: Mapped[Optional[datetime.datetime]] = mapped_column(db.DateTime, nullable=True)
-
-    # Archive status
-    is_archived: Mapped[bool] = mapped_column(db.Boolean, default=False, nullable=False, index=True)
-
-    # Trash (Ruang TPS)
-    is_trashed: Mapped[bool] = mapped_column(db.Boolean, default=False, nullable=False, index=True)
-    trashed_at: Mapped[Optional[datetime.datetime]] = mapped_column(db.DateTime, nullable=True)
-
-    # Folder organization
-    folder_id: Mapped[Optional[int]] = mapped_column(db.Integer, db.ForeignKey('content_folders.id'), nullable=True, index=True)
-    order: Mapped[int] = mapped_column(db.Integer, default=0, nullable=False)
-
-    created_at: Mapped[datetime.datetime] = mapped_column(db.DateTime, default=get_jakarta_now)
-
-    course: Mapped['Course'] = relationship('Course', back_populates='files')
-    folder = relationship('ContentFolder', back_populates='files', foreign_keys=[folder_id])
-
-    def __repr__(self) -> str:
-        return f'<File {self.name}>'
