+++ tests/test_file_upload.py
@@ -12,14 +12,14 @@ class TestFileUploadValidation:
 
     def test_allowed_extension_pdf(self, app):
         """Test PDF file is allowed"""
-        from app.blueprints.courses import allowed_file
+        from app.content.routes import allowed_file
         
         assert allowed_file('document.pdf') is True
         assert allowed_file('file.PDF') is True
 
     def test_allowed_extension_image(self, app):
         """Test image files are allowed"""
-        from app.blueprints.courses import allowed_file
+        from app.content.routes import allowed_file
         
         assert allowed_file('image.png') is True
         assert allowed_file('photo.jpg') is True
@@ -29,7 +29,7 @@ class TestFileUploadValidation:
 
     def test_allowed_extension_document(self, app):
         """Test document files are allowed"""
-        from app.blueprints.courses import allowed_file
+        from app.content.routes import allowed_file
         
         assert allowed_file('document.doc') is True
         assert allowed_file('document.docx') is True
@@ -40,7 +40,7 @@ class TestFileUploadValidation:
 
     def test_disallowed_extension_exe(self, app):
         """Test executable files are disallowed"""
-        from app.blueprints.courses import allowed_file
+        from app.content.routes import allowed_file
         
         assert allowed_file('virus.exe') is False
         assert allowed_file('script.bat') is False
@@ -48,7 +48,7 @@ class TestFileUploadValidation:
 
     def test_disallowed_extension_no_extension(self, app):
         """Test files without extension are disallowed"""
-        from app.blueprints.courses import allowed_file
+        from app.content.routes import allowed_file
         
         assert allowed_file('noextension') is False
         assert allowed_file('file.') is False
